FIX: Update equilibrium solidification printing
The equilibrium solidification final phase fraction printing was
in place before the automatic order/disorder distinguishing in
equilibirum.

Then the final `equilibrium` call outside the solidificaiton
simulation was not using the same point set and could produce
results that were not in agreement with the solidification simulation.

It was also tricky to read the solidification verbose output
correctly, because the phases and fractions at the end of the run
are not sorted in the same order.

Changes in the results JSON files are entirely due to floating
point precision.